Move apiv1 tests into apiv1 folder
